# Simplify topological sorting in `ExprSegmentationSorter::resetLevels`.